Force split on intersection
Rounding errors were causing the intersection to be reported on slightly
different points on the two line segments. This resulted in one of the
segments not getting split.

Now, average the two results for intersection points and force the line
segments to split even if the intersection doesn't seem to actually be
on the segment.