Bug fix: comparison of almost-colinear segments
Because of rounding errors, two segments can be in a situation where the
two endpoints of the first segment appear colinear to the 2nd, but the 2nd's
endpoints do *not* appear colinear to the first.

Previously, the sweep line tree would choke on this because the
compare() method would return the same result for compare(seg1, seg2)
and compare(seg2, seg1).

This fixes that situation by doing a more complete segment comparison to
check for colinearity.