Fix for vertical segments & rounding errors
Because of rounding errors sometimes when an intersection point between
two line segments is calculated it is not exactly on either line.
When this happens with a vertical segment, if the intersection point is
actually to the left of the left endpoint, then the new endpoint should
be the left instead of the right.

To avoid this, this commit implements a constraint that the calculated
intersection point between two line segments is forced to be within
their bounding box overlap - thus keeping vertical segments vertical,
and horizontal segments horizontal.