Invert order of calc of segment intersections
When the input has two already exactly-overlapping segments with a third
that crosses the two, rounding errors can cause slightly different
intersection points to be computed. It appears that calculating the
intersections in this order avoids these rounding errors from creating
disjoint or extraneous segments (at least for this test case, which is
the only one I know about now... there could be other circumstances
which still fail, in which case a bigger fix will be necessary).

Fixes #19